# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface MqttConnectRequest {
  /** MQTT broker URL (e.g., mqtt://broker.example.com:1883) */
  brokerUrl: string;
  username?: string;
  password?: string;
  clientId?: string;
}

export interface MqttSubscribeRequest {
  topic: string;
}

export interface MqttConnectResponse {
  success: boolean;
  message: string;
}

export interface MqttStatusResponse {
  connected: boolean;
  brokerUrl?: string;
  clientId?: string;
  subscribedTopics: string[];
}

export interface ErrorResponse {
  error: string;
}