# lib/api-zod/src/generated/types/mqttConnectRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface MqttConnectRequest {
  /** MQTT broker URL (e.g., mqtt://broker.example.com:1883) */
  brokerUrl: string;
  username?: string;
  password?: string;
  clientId?: string;
}